fix gsa_test, gsa_sample and change imgo_method
