+++ samples/imgo_test.cpp
@@ -1,5 +1,4 @@
 #include <iostream>
-#include <iomanip>
 #include <fstream>
 #include <vector>
 #include <limits>
@@ -9,16 +8,20 @@
 #else
     #include <cmath>
 #endif
+
 #include <imgo.h>
 #include <task.h>
 
-using namespace std;
+using std::ofstream;
+using std::cerr;
+using std::cout;
+using std::endl;
 
 double f1(double x, int j) {
     switch (j) {
         case 1: return exp(-sin(3.0 * x)) - 1.0 / 10.0 * pow(x - 1.0 / 2.0, 2.0) - 1.0;
         case 2: return -13.0 / 6.0 * x + sin(13.0 / 4.0 * (2.0 * x + 5.0)) - 53.0 / 12.0;
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
@@ -26,7 +29,7 @@ double f2(double x, int j) {
     switch (j) {
         case 1: return 1.0 / 20.0 - exp(-2.0 / 5.0 * (x + 5.0)) * sin(4.0 / 5.0 * M_PI * (x + 5.0));
         case 2: return (11.0 * x * x - 10.0 * x + 21.0) / (2.0 * (x * x + 1));
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
@@ -42,7 +45,7 @@ double f4(double x, int j) {
                 sin(2.0 * M_PI * (x - 1.0 / 10.0));
         case 3: return 4.0 * sin(M_PI / 4.0 * x + 1.0 / 20.0) * 
                 pow(pow(sin(M_PI / 2.0 * x + 1.0 / 10.0), 3.0) + pow(cos(M_PI / 2.0 * x + 1.0 / 10.0), 3.0), 2.0);
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
@@ -53,7 +56,7 @@ double f5(double x, int j) {
             71.0 / 10.0 * x * x * x + 79.0 / 20.0 * x * x + x - 1.0 / 10.0);
         case 2: return -14.0 / 125.0 * (3.0 * x - 8.0) * sin(252.0 / 125.0 * (x + 3.0 / 2.0)) - 1.0 / 2.0;
         case 3: return sin(0.423531 * x + 3.13531) + sin(10.0 / 3.0 * (0.423531 * x + 3.13531)) + log(0.423531 * x + 3.13531) + 0.36634 - 0.355766 * x;
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
@@ -62,7 +65,7 @@ double f6(double x, int j) {
         case 1: return 40.0 * (cos(4.0 * x) * (x - sin(x)) * exp( -(x * x) / 2.0));
         case 2: return 2.0 / 25.0 * (x + 4.0) - sin(12.0 / 5.0 * (x + 4.0));
         case 3: return -7.0 / 40.0 * (3.0 * x + 4.0) * sin(63.0 / 20.0 * (x + 4.0));
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
@@ -78,7 +81,7 @@ double f8(double x, int j) {
             return 3.0 / 10.0 - sum;
         case 3: return (-21.0 / 20.0 * x - 13.0 / 8.0) * sin(63.0 / 10.0 * x + 63.0 / 4.0) + 1.0 / 5.0;
         case 4: return cos(7.0 / 4.0 * x + 241.0 / 40.0) - sin(35.0 / 4.0 * x + 241.0 / 8.0) - 5.0;
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
@@ -94,54 +97,53 @@ double f9(double x, int j) {
                 sum += 1.0 / 5.0 * sin((i + 1.0) * x - 1.0) + 2.0;
             }
             return sum;
-        default: return numeric_limits<double>::quiet_NaN();
+        default: return std::numeric_limits<double>::quiet_NaN();
     }
 }
 
 int main() {
     ofstream ofstr("output_data_test.txt");
     if (!ofstr.is_open()) cerr << "File opening error\n";
-    vector<trial> trial_vec;
+    vector<trial_constr> trial_vec;
 
-    double eps = 0.001, r = 3.0, d = 0.0, x_min;
+    double eps = 0.001, r = 3.0, d = 0.0, x;
     int count, Nmax = 1000;
     Stop stop = ACCURACY;
 
-    vector<Task> task_array = { Task(f1, "f1(x)", 1, -2.5, 1.5, 1.05738, eps, Nmax, r, d, stop, 1),
-                                Task(f2, "f2(x)", 1, -5.0, 5.0, 1.016, eps, Nmax, r, d, stop, 1),
-                                Task(f4, "f4(x)", 2, 0.0, 4.0, 2.45956, eps ,Nmax, r, d, stop,  1),
-                                Task(f5, "f5(x)", 2, -1.5, 11.0, 8.85725, eps, Nmax, r, d, stop, 1),
-                                Task(f6, "f6(x)", 2, -4.0, 4.0, 2.32396, eps, Nmax, r, d, stop, 1),
-                                Task(f8, "f8(x)", 3, -2.5, 1.5, -1.12724, eps, Nmax, r, d, stop, 1),
-                                Task(f9, "f9(x)", 3, 0.0, 14.0, 4.0, eps, Nmax, r, d, stop, 1) };
+    vector<task_imgo> task_array = { task_imgo(f1, "f1(x)", 1, -2.5, 1.5, 1.05738, eps, Nmax, r, d, stop),
+                                     task_imgo(f2, "f2(x)", 1, -5.0, 5.0, 1.016, eps, Nmax, r, d, stop),
+                                     task_imgo(f4, "f4(x)", 2, 0.0, 4.0, 2.45956, eps ,Nmax, r, d, stop),
+                                     task_imgo(f5, "f5(x)", 2, -1.5, 11.0, 8.85725, eps, Nmax, r, d, stop),
+                                     task_imgo(f6, "f6(x)", 2, -4.0, 4.0, 2.32396, eps, Nmax, r, d, stop),
+                                     task_imgo(f8, "f8(x)", 3, -2.5, 1.5, -1.12724, eps, Nmax, r, d, stop),
+                                     task_imgo(f9, "f9(x)", 3, 0.0, 14.0, 4.0, eps, Nmax, r, d, stop) };
 
-    imgo_method imgo(nullptr, 0, 0.0, 0.0, r, d, eps);
+    imgo_method imgo(nullptr);
 
     for (int i = 0; i < task_array.size(); i++) {
-        imgo.setFunc(task_array[i].f);
+        imgo.setF(task_array[i].f);
         imgo.setM(task_array[i].m);
-        imgo.setAB(task_array[i].A, task_array[i].B);
+        imgo.setAB(task_array[i].A[0], task_array[i].B[0]);
         imgo.setEps(task_array[i].eps);
         imgo.setNmax(task_array[i].Nmax);
         imgo.setR(task_array[i].r);
         imgo.setD(task_array[i].d);
-        imgo.setNmax(task_array[i].Nmax);
 
-        x_min = imgo.solve(count, task_array[i].stop);
+        imgo.solve(count, x, task_array[i].stop);
 
         cout << "Function: " << task_array[i].name << endl;
         cout << "[a; b] = [" << task_array[i].A[0] << "; " << task_array[i].B[0] << "]"<< endl;
-        cout << "X* = " << setprecision(12) << task_array[i].X_opt[0] << endl;
-        cout << "X = " << setprecision(12) << x_min << endl;
-        cout << "|X* - X| = " << abs(task_array[i].X_opt[0] - x_min) << endl;
+        cout << "X* = " << task_array[i].X_opt[0] << endl;
+        cout << "X = " << x << endl;
+        cout << "|X* - X| = " << abs(task_array[i].X_opt[0] - x) << endl;
         cout << "|f(X*) - f(X)| = " << abs(task_array[i].f(task_array[i].X_opt[0], task_array[i].m + 1) - 
-                                           task_array[i].f(x_min, task_array[i].m + 1)) << endl;
+                                           task_array[i].f(x, task_array[i].m + 1)) << endl;
         cout << "Number of trials = " << count << endl;
         cout << endl;
 
         imgo.getTrialPoints(trial_vec);
         ofstr << task_array[i].A[0] << " " << task_array[i].B[0] << " " << task_array[i].m 
-              << " " << x_min << " " << task_array[i].X_opt[0] << endl;
+              << " " << x << " " << task_array[i].X_opt[0] << endl;
         for (int j = 0; j < trial_vec.size(); j++) {
             ofstr << trial_vec[j].x << " " << trial_vec[j].z << endl;
         }
