+++ samples/operational_characteristics.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+/* #include <iostream>
 #include <fstream>
 #include <limits>
 #include <Hill/HillProblemFamily.hpp>
@@ -142,4 +142,4 @@ int main() {
     cin.get();
 #endif
 	return 0;
-}
+} */
