+++ samples/peano_sample.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+/* #include <iostream>
 #include <fstream>
 #include <cstdlib>
 #include <cmath>
@@ -124,3 +124,4 @@ int main() {
 
     return 0;
 }
+ */
