+++ samples/peano_test.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+/* #include <iostream>
 #include <fstream> 
 #include <cmath>
 #include <Grishagin/grishagin_function.hpp>
@@ -123,3 +123,4 @@ int main() {
 
     return 0;
 }
+ */
