--- src/task.cpp
@@ -1,1 +1,0 @@
-#include <task.h>
